requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.6.2            # via -r requirements.in
arrow==0.14.7             # via -r requirements.in, ics
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
chardet==3.0.4            # via aiohttp
ics==0.7                  # via -r requirements.in
idna==2.9                 # via yarl
lxml==4.5.0               # via -r requirements.in
multidict==4.7.5          # via aiohttp, yarl
python-dateutil==2.8.1    # via arrow, ics
python-dotenv==0.12.0     # via -r requirements.in
six==1.14.0               # via ics, python-dateutil
tatsu==4.4.0              # via ics
yarl==1.4.2               # via aiohttp
